Added files via upload
These are the source files for my demo project that uses Selenium WebDriver calls, implemented in Java and running in a JUnit framework, to send and receive mail in Google Gmail. The readme.txt file describes this project in some detail. It is intended to demonstrate my level of skill in the use of Selenium WebDriver as of the time I created it (2014). This project was created in the Kepler version of Eclipse. In theory, if one were to extract the contents of GmailSeleniumDemo.zip to a new project in Eclipse, and then add Selenium libraries as documented in howtobuildseleniumstuff.doc, you might be able to build this in Eclipse. Even if that is successful, it is possible that this may not run as intended due to changes in Google, Gmail or other dependencies I am not aware of.